Move Regex Shared/Exclusive Reference to own file
